research/AMP-Rag-Extension/pyproject.toml
[tool.poetry]
name = "amp-rag-extension"
version = "0.1.0"
description = ""
authors = ["yunyoseob <imyoseob@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
scrapy = "^2.11.2"
pyyaml = "^6.0.1"
beautifulsoup4 = "^4.12.3"
lxml = "^5.2.2"
apscheduler = "^3.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
